+++ src/Persistence/Schema/MigrationList.php
@@ -18,11 +18,13 @@ use EnergyCRM\Persistence\Schema\Migrations\AddContractAddresses;
 use EnergyCRM\Persistence\Schema\Migrations\AddContractListIndexes;
 use EnergyCRM\Persistence\Schema\Migrations\AddCustomerAfmIndex;
 use EnergyCRM\Persistence\Schema\Migrations\AddForeignKeys;
+use EnergyCRM\Persistence\Schema\Migrations\AddProgramCodeColumn;
 use EnergyCRM\Persistence\Schema\Migrations\EnsureInnoDb;
 use EnergyCRM\Persistence\Schema\Migrations\EnsureLegacyColumns;
 use EnergyCRM\Persistence\Schema\Migrations\FixProviderEnergyTypes;
 use EnergyCRM\Persistence\Schema\Migrations\MoveMeterAddressOutOfExtras;
 use EnergyCRM\Persistence\Schema\Migrations\SeedMobilePrograms;
+use EnergyCRM\Persistence\Schema\Migrations\SeedOrizonPlans;
 use EnergyCRM\Persistence\Schema\Migrations\PurgeOrphanFiles;
 use EnergyCRM\Persistence\Schema\Migrations\WidenEncryptedColumns;
 
@@ -47,6 +49,8 @@ final class MigrationList
             new SeedMobilePrograms(),
             new AddCustomerAfmIndex(),
             new WidenEncryptedColumns(),
+            new AddProgramCodeColumn(),
+            new SeedOrizonPlans(),
         ];
     }
 }